Optionally do not include Montagu test data

+++ proxy/scripts/run-dependencies.sh
@@ -34,26 +34,24 @@ migrate_image=$REGISTRY/montagu-migrate:master
 docker pull $migrate_image
 docker run --network=montagu_proxy $migrate_image
 
-# Generate test data, including reports, if 'data' present as first param
-#if [ "$1" = "data" ]; then
+# Generate test data if 'data' present as first param
+if [ "$1" = "data" ]; then
   test_data_image=$REGISTRY/montagu-generate-test-data:master
   docker pull $test_data_image
   docker run --rm --network=montagu_proxy $test_data_image
+fi
 
-  rm demo -rf
-  rm git -rf
-  # Generate report test data
-    docker pull $REGISTRY/orderly:master
-    docker run --rm \
-      --entrypoint create_orderly_demo.sh \
-      -u $UID \
-      -v $PWD:/orderly \
-      -w "/orderly" \
-      $REGISTRY/orderly:master \
-      "."
-
-
-#fi
+# Always generate report test database
+rm demo -rf
+rm git -rf
+docker pull $REGISTRY/orderly:master
+docker run --rm \
+  --entrypoint create_orderly_demo.sh \
+  -u $UID \
+  -v $PWD:/orderly \
+  -w "/orderly" \
+  $REGISTRY/orderly:master \
+  "."
 
 # Copy the demo dbdocker  file to top level
 docker cp $PWD/demo/orderly.sqlite montagu_orderly_web_web_1:/orderly/orderly.sqlite
